+++ CorePayments/src/main/java/com/paypal/android/corepayments/graphql/GraphQLRequest.kt
@@ -4,13 +4,12 @@ import androidx.annotation.RestrictTo
 import kotlinx.serialization.InternalSerializationApi
 import kotlinx.serialization.Serializable
 
-/**
- * @suppress
- */
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-@InternalSerializationApi
+@OptIn(InternalSerializationApi::class)
 @Serializable
-data class GraphQLError(
-    val message: String,
-    val extensions: List<GraphQLExtension>? = null
+data class GraphQLRequest<V>(
+    val query: String,
+    val variables: V? = null,
+    @kotlinx.serialization.Transient
+    val operationName: String? = null
 )