+++ CorePayments/src/main/java/com/paypal/android/corepayments/graphql/GraphQLResponse.kt
@@ -3,14 +3,13 @@ package com.paypal.android.corepayments.graphql
 import androidx.annotation.RestrictTo
 import kotlinx.serialization.InternalSerializationApi
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.JsonObject
 
-/**
- * @suppress
- */
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 @InternalSerializationApi
 @Serializable
-data class GraphQLError(
-    val message: String,
-    val extensions: List<GraphQLExtension>? = null
+data class GraphQLResponse<out T> @OptIn(InternalSerializationApi::class) constructor(
+    val data: T? = null,
+    val extensions: JsonObject? = null,
+    val errors: List<GraphQLError>? = null,
 )